Reformat entire command line procedure
