--- eng/apiGenerator/DmsOpenApiGenerator/DmsOpenApiGenerator.sln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.12.35707.178 d17.12
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DmsOpenApiGenerator", "DmsOpenApiGenerator.csproj", "{42DCCDB3-F90B-4C8D-B692-2757B49C7ABF}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DmsOpenApiGenerator.Tests", "..\DmsOpenApiGenerator.Tests\DmsOpenApiGenerator.Tests.csproj", "{8FB8461F-36D1-4DA4-89C5-4DEAD74B140F}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{42DCCDB3-F90B-4C8D-B692-2757B49C7ABF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{42DCCDB3-F90B-4C8D-B692-2757B49C7ABF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{42DCCDB3-F90B-4C8D-B692-2757B49C7ABF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{42DCCDB3-F90B-4C8D-B692-2757B49C7ABF}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8FB8461F-36D1-4DA4-89C5-4DEAD74B140F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8FB8461F-36D1-4DA4-89C5-4DEAD74B140F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8FB8461F-36D1-4DA4-89C5-4DEAD74B140F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8FB8461F-36D1-4DA4-89C5-4DEAD74B140F}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
