+++ src/dms/clis/EdFi.DataManagementService.OpenApiGenerator.Tests.Unit/EdFi.DataManagementService.OpenApiGenerator.Tests.Unit.csproj
@@ -1,0 +1,30 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net8.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+
+    <IsPackable>false</IsPackable>
+    <IsTestProject>true</IsTestProject>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="coverlet.collector"/>
+    <PackageReference Include="Microsoft.NET.Test.Sdk" />
+    <PackageReference Include="NUnit" />
+    <PackageReference Include="NUnit.Analyzers" />
+    <PackageReference Include="NUnit3TestAdapter" />
+    <PackageReference Include="FluentAssertions" />
+    <PackageReference Include="FakeItEasy"  />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Using Include="NUnit.Framework" />
+  </ItemGroup>
+
+    <ItemGroup>
+        <ProjectReference Include="..\EdFi.DataManagementService.OpenApiGenerator\EdFi.DataManagementService.OpenApiGenerator.csproj" />
+    </ItemGroup>
+
+</Project>
